Separate route tree collection builder
